+++ src/__test__/components/ChildComponent.nuxt.spec.ts
@@ -1,12 +1,12 @@
 import { describe, expect, test } from 'vitest';
-import ChildComponent from '@/components/props/ChildComponent.vue';
+import ChildComponent from '@/components/ChildComponent.vue';
 import { mountSuspendedComponent } from '@/helpers/test';
 
-describe('ChildComponent.vue', () => {
-  describe('Propsのテスト', () => {
+describe('src/components/ChildComponent.vue', () => {
+  describe('Propsテスト', () => {
     test('親コンポーネントから受け取ったPropsの値が適切に表示されるか', async () => {
       const props = {
-        text: 'Test Props',
+        textContent: 'Test Props',
       };
       const wrapper = await mountSuspendedComponent(ChildComponent, { props });
