--- src/components/mount/ChildComponent.vue
@@ -1,3 +1,0 @@
-<template>
-  <div id="child-component">Child Component</div>
-</template>
