--- src/components/mount/Mount.vue
@@ -1,3 +1,0 @@
-<template>
-  <div id="mount">mount test</div>
-</template>
