--- src/components/mount/ParentComponent.vue
@@ -1,8 +1,0 @@
-<script setup lang="ts">
-import ChildComponent from '@/components/mount/ChildComponent.vue';
-</script>
-<template>
-  <div id="parent-component">
-    <ChildComponent />
-  </div>
-</template>
