--- src/components/props/ParentComponent.vue
@@ -1,8 +1,0 @@
-<script setup lang="ts">
-import ChildComponent from '@/components/props/ChildComponent.vue';
-</script>
-<template>
-  <div id="parent-component">
-    <ChildComponent text="Hello World" />
-  </div>
-</template>
