+++ src/pages/index.vue
@@ -1,10 +1,3 @@
 <template>
-  <ul>
-    <li>
-      <NuxtLink to="/mount">Mount</NuxtLink>
-    </li>
-    <li>
-      <NuxtLink to="/props">Props</NuxtLink>
-    </li>
-  </ul>
+  <div>テスト</div>
 </template>
