--- src/pages/mount/index.vue
@@ -1,10 +1,0 @@
-<script setup lang="ts">
-import Mount from '@/components/mount/Mount.vue';
-import ParentComponent from '@/components/mount/ParentComponent.vue';
-</script>
-<template>
-  <section>
-    <Mount />
-    <ParentComponent />
-  </section>
-</template>
