--- src/pages/props/index.vue
@@ -1,6 +1,0 @@
-<script setup lang="ts">
-import ParentComponent from '@/components/props/ParentComponent.vue';
-</script>
-<template>
-  <ParentComponent />
-</template>
